# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem Manajemen Karyawan API",
        "description": "Dokumentasi API untuk sistem manajemen karyawan, absensi, dan penggajian",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/karyawan": {
            "get": {
                "tags": [
                    "Karyawan"
                ],
                "summary": "Mendapatkan daftar semua karyawan",
                "operationId": "974f1d3446b8f4b830a6194f05289a9a",
                "responses": {
                    "200": {
                        "description": "Data karyawan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data karyawan berhasil diambil."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_karyawan": {
                                                        "type": "string",
                                                        "example": "K001"
                                                    },
                                                    "nama": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "jabatan": {
                                                        "type": "string",
                                                        "example": "Staff IT"
                                                    },
                                                    "gaji_pokok": {
                                                        "type": "number",
                                                        "example": 5000000
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Karyawan"
                ],
                "summary": "Menambahkan karyawan baru",
                "operationId": "a9c07227b0b719b7742e937e517c95e4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_karyawan": {
                                        "type": "string",
                                        "example": "K001"
                                    },
                                    "nama": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "jabatan": {
                                        "type": "string",
                                        "example": "Staff IT"
                                    },
                                    "gaji_pokok": {
                                        "type": "number",
                                        "example": 5000000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Karyawan berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data karyawan berhasil dibuat"
                                        },
                                        "data": {
                                            "properties": {
                                                "id_karyawan": {
                                                    "type": "string",
                                                    "example": "K001"
                                                },
                                                "nama": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "jabatan": {
                                                    "type": "string",
                                                    "example": "Staff IT"
                                                },
                                                "gaji_pokok": {
                                                    "type": "number",
                                                    "example": 5000000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Gagal membuat data karyawan"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "The id karyawan field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/karyawan/{id}/absensi": {
            "get": {
                "tags": [
                    "Karyawan"
                ],
                "summary": "Mendapatkan data absensi karyawan berdasarkan ID",
                "operationId": "80785ef0623798b52d516bc8654ad5ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID karyawan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data absensi karyawan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data absensi karyawan berhasil diambil."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "id_karyawan": {
                                                        "type": "string",
                                                        "example": "K001"
                                                    },
                                                    "tanggal": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2023-05-01"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "Hadir"
                                                    },
                                                    "keterangan": {
                                                        "type": "string",
                                                        "example": "Masuk tepat waktu"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Karyawan tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data karyawan tidak ditemukan."
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Masukkan API key dengan format Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}